Crearea aplicatiei python
